Wave adjustments/enemy prototype
-adjustments to the echolocation tool to make it match the design doc/fix a bug
-unfinished enemy controller script, ignore it

+++ Game Club Club Jam Jam Game/Assets/echo_location/echo_wave_visuals.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class echo_wave_visuals : MonoBehaviour
 {
     [SerializeField] LineRenderer lineRenderer;
-    [SerializeField] int expansion_rate;
+    [SerializeField] float expansion_rate;
     [SerializeField] float max_radius;
     [SerializeField] int segments;
     [SerializeField] public Vector2 direction;
